requirements.txt
pybit>=5.7.0
python-telegram-bot>=20.7
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0

